services/capture/pyproject.toml
[project]
name = "capture"
version = "0.1.0"
description = "User input capture service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.65",
    "fastapi[standard]>=0.121.0",
    "psycopg-pool>=3.2.7",
    "psycopg[binary]>=3.2.12",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
